C >>> 1 instead of c >> 1 in polynomialGaussSamplerIII from Sample.java
PARAM_XI 12.0096 --> 12 in Parameter.java

+++ qTESLA/src/qTESLA/QTESLA.java
@@ -1050,7 +1050,7 @@ public class QTESLA {
 		
 		for (int i = 0; i < w; i++) {
 			
-			for (short j = 0; j < limit - 1; j++) {
+			for (int j = 0; j < limit - 1; j++) {
 				
 				mask		= (short) ((list[j + 1] - list[j]) >> 15);
 				temporary	= (short) ((list[j + 1] & mask) | (list[j] & (~ mask)));
